requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cssselect2==0.6.0
Django==4.0.4
djangorestframework==3.13.1
easy-thumbnails==2.8.1
easy-thumbnails-rest==1.1.4
lxml==4.8.0
Pillow==9.1.1
python-decouple==3.6
pytz==2022.1
reportlab==3.6.9
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
webencodings==0.5.1
